feat(auth): refactor token extraction and decoding
